added the authorization in controller and modified program.cs

+++ JobManagement.Repositories/IUserRepository.cs
@@ -9,7 +9,7 @@ namespace JobManagement.Repositories
 {
     public interface IUserRepository
     {
-        public Task<user> GetUserById(int userId);
+        public Task<user> GetUserById(long userId);
         public Task<user> GetUserByEmail(string email);
         public Task<user> AddUser(user user);
         public Task<user> UpdateUser(user user);
